+++ services/orchest-api/app/app/utils.py
@@ -1,7 +1,10 @@
 import logging
 import time
+import uuid
+from collections import ChainMap
+from copy import deepcopy
 from datetime import datetime
-from typing import Container, Dict, Iterable, Optional, Union
+from typing import Container, Dict, Iterable, List, Optional, Tuple, Union
 
 from celery.utils.log import get_task_logger
 from flask import current_app
@@ -9,10 +12,12 @@ from flask_restx import Model, Namespace
 from flask_sqlalchemy import Pagination
 from kubernetes import client as k8s_client
 from sqlalchemy import or_, text
+from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.orm import query, undefer
 
 import app.models as models
 from _orchest.internals import config as _config
+from _orchest.internals import errors as _errors
 from app import errors as self_errors
 from app import schema
 from app.connections import db, k8s_core_api
@@ -220,3 +225,299 @@ def get_jupyter_server_image_to_use() -> str:
         return f"{registry_ip}/{_config.JUPYTER_IMAGE_NAME}:latest"
     else:
         return f"orchest/jupyter-server:{CONFIG_CLASS.ORCHEST_VERSION}"
+
+
+class GlobalOrchestConfig:
+    _cloud = _config.CLOUD
+
+    # Defines default values for all supported configuration options.
+    _config_values = {
+        "MAX_JOB_RUNS_PARALLELISM": {
+            "default": 1,
+            "type": int,
+            "requires-restart": True,
+            "condition": lambda x: 0 < x <= 25,
+            "condition-msg": "within the range [1, 25]",
+        },
+        "MAX_INTERACTIVE_RUNS_PARALLELISM": {
+            "default": 1,
+            "type": int,
+            "requires-restart": True,
+            "condition": lambda x: 0 < x <= 25,
+            "condition-msg": "within the range [1, 25]",
+        },
+        "AUTH_ENABLED": {
+            "default": False,
+            "type": bool,
+            "requires-restart": True,
+            "condition": None,
+        },
+        "TELEMETRY_DISABLED": {
+            "default": False,
+            "type": bool,
+            "requires-restart": True,
+            "condition": None,
+        },
+        "TELEMETRY_UUID": {
+            "default": str(uuid.uuid4()),
+            "type": str,
+            "requires-restart": True,
+            "condition": None,
+        },
+        "INTERCOM_USER_EMAIL": {
+            "default": "johndoe@example.org",
+            "type": str,
+            "requires-restart": True,
+            "condition": None,
+        },
+    }
+    _cloud_unmodifiable_config_opts = [
+        "TELEMETRY_UUID",
+        "TELEMETRY_DISABLED",
+        "AUTH_ENABLED",
+        "INTERCOM_USER_EMAIL",
+    ]
+
+    def __init__(self) -> None:
+        """Manages the global user config.
+
+        Uses a collections.ChainMap under the hood to provide fallback
+        to default values where needed. And when running with `--cloud`,
+        it won't allow you to update config values of the keys defined
+        in `self._cloud_unmodifiable_config_opts`.
+
+        Example:
+            >>> config = GlobalOrchestConfig()
+            >>> # Set the current config to a new one.
+            >>> config.set(new_config)
+            >>> # Save the updated (and automatically validated) config
+            >>> # to disk.
+            >>> requires_orchest_restart = config.save(flask_app=app)
+            >>> # Just an example output.
+            >>> requres_orchest_restart
+            ... ["MAX_INTERACTIVE_RUNS_PARALLELISM"]
+
+        """
+        unmodifiable_config, current_config = self._get_current_configs()
+        defaults = {k: val["default"] for k, val in self._config_values.items()}
+
+        self._values = ChainMap(unmodifiable_config, current_config, defaults)
+
+    def as_dict(self) -> dict:
+        # Flatten into regular dictionary.
+        return dict(self._values)
+
+    def save(self, flask_app=None) -> Optional[List[str]]:
+        """Saves the state to the database.
+
+        Args:
+            flask_app (flask.Flask): Uses the `flask_app.config` to
+                determine whether Orchest needs to be restarted for the
+                global config changes to take effect.
+
+        Returns:
+            * `None` if no `flask_app` is given.
+            * List of changed config options that require an Orchest
+              restart to take effect.
+            * Empty list otherwise.
+
+        """
+        settings_as_dict = self.as_dict()
+
+        # Upsert entries.
+        stmt = insert(models.Setting).values(
+            [dict(name=k, value={"value": v}) for k, v in settings_as_dict.items()]
+        )
+        stmt = stmt.on_conflict_do_update(
+            index_elements=[models.Setting.name], set_=dict(value=stmt.excluded.value)
+        )
+        db.session.execute(stmt)
+
+        # Delete settings that are not part of the new configuration.
+        models.Setting.query.filter(
+            models.Setting.name.not_in(list(settings_as_dict.keys()))
+        ).delete()
+
+        db.session.commit()
+
+        if flask_app is None:
+            return
+
+        return self._changes_require_restart(flask_app, settings_as_dict)
+
+    def update(self, d: dict) -> None:
+        """Updates the current config values.
+
+        Under the hood it just calls `dict.update` on the current config
+        dict.
+
+        Raises:
+            TypeError: The values of the dictionary that correspond to
+                supported config values have incorrect types.
+            ValueError: The values of the dictionary that correspond to
+                supported config values have incorrect values. E.g.
+                maximum parallelism has to be greater or equal to one.
+
+        """
+        try:
+            self._validate_dict(d)
+        except (TypeError, ValueError) as e:
+            current_app.logger.error(
+                "Tried to update global Orchest config with incorrect types or values."
+            )
+            raise e
+        else:
+            self._values.maps[1].update(d)
+
+    def set(self, d: dict) -> None:
+        """Overwrites the current config with the given dict.
+
+        Raises:
+            TypeError: The values of the dictionary that correspond to
+                supported config values have incorrect types.
+            ValueError: The values of the dictionary that correspond to
+                supported config values have incorrect values. E.g.
+                maximum parallelism has to be greater or equal to one.
+
+        """
+        try:
+            self._validate_dict(d)
+        except (TypeError, ValueError) as e:
+            current_app.logger.error(
+                "Tried to update global Orchest config with incorrect types or values."
+            )
+            raise e
+        else:
+            self._values.maps[1] = d
+
+    def __getitem__(self, key):
+        return self._values[key]
+
+    def _changes_require_restart(self, flask_app, new: dict) -> List[str]:
+        """Do config changes require an Orchest restart.
+
+        Compares the Orchest global config values in the flask app to
+        the `new` values and determines whether the changes require a
+        restart of the Orchest application.
+
+        Returns:
+            A list of strings representing the changed configuration
+            options that require a restart of Orchest to take effect.
+
+        """
+        res = []
+        for k, val in self._config_values.items():
+            if not val["requires-restart"]:
+                continue
+
+            # Changes to unmodifiable config options won't take effect
+            # anyways and so they should not account towards requiring
+            # a restart yes or no.
+            if self._cloud and k in self._cloud_unmodifiable_config_opts:
+                continue
+
+            old_val = flask_app.config[k]
+            if new.get(k) is not None and new[k] != old_val:
+                res.append(k)
+
+        return res
+
+    def _validate_dict(self, d: dict, migrate=False) -> None:
+        """Validates the types and values of the values of the dict.
+
+        Validates whether the types of the values of the given dict
+        equal the types of the respective key's values of the
+        `self._config_values` and additional key specific rules are
+        satisfied, e.g. parallelism > 0.
+
+        Args:
+            d: The dictionary to validate the types and values of.
+            migrate: If `True`, then the options for which the type
+                and/or value are invalidated get assigned their default
+                value. However, `self._cloud_unmodifiable_config_opts`
+                are never migrated if `self._cloud==True` as that could
+                cause authentication to get disabled.
+
+        Note:
+            Keys in the given dict that are not in the
+            `self._config_values` are not checked.
+
+        """
+        for k, val in self._config_values.items():
+            try:
+                given_val = d[k]
+            except KeyError:
+                # We let it pass silently because it won't break the
+                # application in any way as we will later fall back on
+                # default values.
+                current_app.logger.debug(
+                    f"Missing value for required config option: {k}."
+                )
+                continue
+
+            if type(given_val) is not val["type"]:
+                not_allowed_to_migrate = (
+                    self._cloud and k in self._cloud_unmodifiable_config_opts
+                )
+                if not migrate or not_allowed_to_migrate:
+                    given_val_type = type(given_val).__name__
+                    correct_val_type = val["type"].__name__
+                    raise TypeError(
+                        f'{k} has to be a "{correct_val_type}" but "{given_val_type}"'
+                        " was given."
+                    )
+
+                d[k] = val["default"]
+
+            if val["condition"] is not None and not val["condition"].__call__(
+                given_val
+            ):
+                not_allowed_to_migrate = (
+                    self._cloud and k in self._cloud_unmodifiable_config_opts
+                )
+                if not migrate or not_allowed_to_migrate:
+                    raise ValueError(f"{k} has to be {val['condition-msg']}.")
+
+                d[k] = val["default"]
+
+    def _get_current_configs(self) -> Tuple[dict, dict]:
+        """Gets the dicts needed to initialize this class.
+
+        Returns:
+            (unmodifiable_config, current_config): The first being
+                populated in case `self._cloud==True` and taking the
+                values of the respective `current_config` values.
+
+        """
+        current_config = self._fetch_settings_from_db()
+
+        try:
+            # Make sure invalid values are migrated to default values,
+            # because the application can not start with invalid values.
+            self._validate_dict(current_config, migrate=True)
+        except (TypeError, ValueError):
+            raise _errors.CorruptedFileError(
+                f'Option(s) defined in the global user config ("{self._path}") has'
+                + " incorrect type and/or value."
+            )
+
+        unmodifiable_config = {}
+        if self._cloud:
+            for k in self._cloud_unmodifiable_config_opts:
+                try:
+                    unmodifiable_config[k] = deepcopy(current_config[k])
+                except KeyError:
+                    # Fall back on default values.
+                    ...
+
+        return unmodifiable_config, current_config
+
+    def _fetch_settings_from_db(self) -> dict:
+        """Fetches the settings from the database."""
+
+        stored_settings = models.Setting.query.all()
+        settings = {}
+        for setting in stored_settings:
+            settings[setting.name] = setting.value["value"]
+
+        return settings
